Move Battle to battle
