+++ core/Factory.py
@@ -1,4 +1,4 @@
-from Battle import BattleCog, ToonAttack
+from battle.Battle import BattleCog, ToonAttack
 from Cog import Cog
 from core.Entity import BattleEntity, Entity
 from Gag import Gag, get_gag_min_max_exp
