+++ tests/battle/test_battle_cog_creation.py
@@ -1,4 +1,4 @@
-from Battle import BattleCog
+from battle.Battle import BattleCog
 from Cog import Cog
 from core.Entity import BattleEntity, Entity
 from core.Factory import BattleCogFactory, BattleEntityFactory, CogFactory
